feat(PAV-7): conexoes como secao dentro do perfil
Move a gestao de conexoes sociais para uma secao no ProfileTab,
seguindo o design system dos demais cards do perfil, e aposenta a
pagina/rota standalone /perfil/conexoes.

- ConnectionsForm: nova secao no perfil (getConnections/connect/
  disconnect, guarda de ultimo metodo de login, feedback via
  ?linked=/?error=)
- ProfileTab: renderiza ConnectionsForm como terceira secao
- AppRoutes: /perfil/conexoes vira redirect para /perfil;
  ConnectionsPage removida
- auth.controller: redirect do vinculo OAuth aponta para /perfil
  (+ teste atualizado)
- vite.config: navegacoes de pagina do OAuth caem no SPA em vez do
  proxy de API
