+++ frontend/src/domains/new_dashboard/components/profile/ProfileTab.tsx
@@ -1,4 +1,5 @@
 import type { SearchPreferences, UserProfile } from "../../types";
+import { ConnectionsForm } from "./ConnectionsForm";
 import { PreferencesForm } from "./PreferencesForm";
 import { ProfileForm } from "./ProfileForm";
 
@@ -37,6 +38,7 @@ export function ProfileTab({
         isSaving={isSavingPreferences}
         onSave={onSavePreferences}
       />
+      <ConnectionsForm />
     </div>
   );
 }
